08-mensajeria-async/02-python-a-python2/subscriber/requirements.txt
pika
python-dotenv
Flask
flasgger
